# contracts/openapi.json
{
  "components": {
    "schemas": {
      "Business": {
        "properties": {
          "enterprise_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enterprise Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "receiving_processes": {
            "items": {
              "type": "string"
            },
            "title": "Receiving Processes",
            "type": "array"
          },
          "roles": {
            "items": {
              "enum": [
                "buyer",
                "seller"
              ],
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "roles",
          "receiving_processes",
          "location"
        ],
        "title": "Business",
        "type": "object"
      },
      "Costs": {
        "properties": {
          "buyer_total_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Buyer Total Paise",
            "type": "integer"
          },
          "freight_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Freight Paise",
            "type": "integer"
          },
          "material_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Material Paise",
            "type": "integer"
          },
          "seller_receives_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Seller Receives Paise",
            "type": "integer"
          }
        },
        "required": [
          "material_paise",
          "freight_paise",
          "buyer_total_paise",
          "seller_receives_paise"
        ],
        "title": "Costs",
        "type": "object"
      },
      "Deal": {
        "properties": {
          "buyer": {
            "$ref": "#/components/schemas/Business"
          },
          "costs": {
            "$ref": "#/components/schemas/Costs"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "intended_use": {
            "title": "Intended Use",
            "type": "string"
          },
          "listing_id": {
            "title": "Listing Id",
            "type": "string"
          },
          "material_id": {
            "title": "Material Id",
            "type": "string"
          },
          "negotiation_id": {
            "title": "Negotiation Id",
            "type": "string"
          },
          "quantity_kg": {
            "description": "kilograms",
            "minimum": 0.0,
            "title": "Quantity Kg",
            "type": "integer"
          },
          "requirement_id": {
            "title": "Requirement Id",
            "type": "string"
          },
          "seller": {
            "$ref": "#/components/schemas/Business"
          },
          "status": {
            "enum": [
              "agreed",
              "pickup_scheduled",
              "collected",
              "delivered",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "transport_option": {
            "$ref": "#/components/schemas/TransportOption"
          },
          "unit_price_paise_per_tonne": {
            "description": "INR paise per tonne",
            "minimum": 0.0,
            "title": "Unit Price Paise Per Tonne",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "negotiation_id",
          "requirement_id",
          "listing_id",
          "seller",
          "buyer",
          "material_id",
          "intended_use",
          "quantity_kg",
          "unit_price_paise_per_tonne",
          "costs",
          "transport_option",
          "status",
          "created_at"
        ],
        "title": "Deal",
        "type": "object"
      },
      "DealStatusUpdate": {
        "properties": {
          "status": {
            "enum": [
              "pickup_scheduled",
              "collected",
              "delivered",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "DealStatusUpdate",
        "type": "object"
      },
      "DistrictRef": {
        "properties": {
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "lat",
          "lon"
        ],
        "title": "DistrictRef",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Location": {
        "properties": {
          "district": {
            "title": "District",
            "type": "string"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "precision": {
            "default": "district",
            "enum": [
              "district",
              "confirmed_site"
            ],
            "title": "Precision",
            "type": "string"
          }
        },
        "required": [
          "district"
        ],
        "title": "Location",
        "type": "object"
      },
...
      "MaterialRef": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "MaterialRef",
        "type": "object"
      },
      "MeResponse": {
        "properties": {
          "business": {
            "$ref": "#/components/schemas/Business"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "business"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "Page_Deal_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Deal"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[Deal]",
        "type": "object"
      },
...
      "ReceivingProcessRef": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "material_ids": {
            "items": {
              "type": "string"
            },
            "title": "Material Ids",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "material_ids"
        ],
        "title": "ReceivingProcessRef",
        "type": "object"
      },
      "ReferenceResponse": {
        "properties": {
          "districts": {
            "items": {
              "$ref": "#/components/schemas/DistrictRef"
            },
            "title": "Districts",
            "type": "array"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/MaterialRef"
            },
            "title": "Materials",
            "type": "array"
          },
          "receiving_processes": {
            "items": {
              "$ref": "#/components/schemas/ReceivingProcessRef"
            },
            "title": "Receiving Processes",
            "type": "array"
          }
        },
        "required": [
          "materials",
          "districts",
          "receiving_processes"
        ],
        "title": "ReferenceResponse",
        "type": "object"
      },
...
      "TransportOption": {
        "properties": {
          "capacity_kg": {
            "description": "kilograms",
            "minimum": 0.0,
            "title": "Capacity Kg",
            "type": "integer"
          },
          "delivery_at": {
            "format": "date-time",
            "title": "Delivery At",
            "type": "string"
          },
          "distance_basis": {
            "default": "unknown",
            "enum": [
              "road",
              "district_straight_line",
              "unknown"
            ],
            "title": "Distance Basis",
            "type": "string"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "freight_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Freight Paise",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "listing_id": {
            "title": "Listing Id",
            "type": "string"
          },
          "pickup_at": {
            "format": "date-time",
            "title": "Pickup At",
            "type": "string"
          },
          "requirement_id": {
            "title": "Requirement Id",
            "type": "string"
          },
          "source": {
            "enum": [
              "entered_quote",
              "configured_estimate"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "id",
          "listing_id",
          "requirement_id",
          "label",
          "freight_paise",
          "capacity_kg",
          "pickup_at",
          "delivery_at",
          "expires_at",
          "source"
        ],
        "title": "TransportOption",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AccessToken": {
        "description": "AUTH_MODE=dev: paste `dev:<email>` (for example `dev:buyer1@demo.bitnbuild.local`). AUTH_MODE=supabase: paste a Supabase access token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "B2B exchange for secondary industrial materials. Prices are entered by participants; freight charges are entered or configured estimates. No market price feed, benchmark, or carrier booking is involved.",
    "title": "Karnataka Industrial Byproduct Exchange",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/deals/{deal_id}": {
      "get": {
        "operationId": "get_deal_api_v1_deals__deal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "title": "Deal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deal"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Get Deal",
        "tags": [
          "deals"
        ]
      }
    },
    "/api/v1/deals/{deal_id}/status": {
      "patch": {
        "operationId": "update_deal_status_api_v1_deals__deal_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "deal_id",
            "required": true,
            "schema": {
              "title": "Deal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DealStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deal"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Update Deal Status",
        "tags": [
          "deals"
        ]
      }
    },
...
    "/api/v1/me": {
      "get": {
        "operationId": "me_api_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Me",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/me/deals": {
      "get": {
        "operationId": "my_deals_api_v1_me_deals_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Deal_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "My Deals",
        "tags": [
          "deals"
        ]
      }
    },
...
    "/api/v1/reference": {
      "get": {
        "operationId": "reference_api_v1_reference_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Reference",
        "tags": [
          "reference"
        ]
      }
    },
...
  }
}